fix: refactor Route paths and GalleryPage layout
