+++ src/views/CardDetails/CardDetails.css
@@ -1,4 +1,4 @@
 .card-details {
-  min-width: 200px;
+  min-width: 400px;
   background-color: blue;
 }
